Обновление usb4java до 0.7 версии. Ранее обнаружена ошибка в Windows 7 связанная с библиотекой и последней версией jre 1.8. Новая версия библиотеки решила проблему, но немного изменился API. Учтено что на Windows 7 не работает .net 5 и пользователю для обновления будет предложено скачать инсталлер с сайта, уже при входе в аккаунт, а не в стартере.

+++ mair_app/src/main/java/ru/biomedis/biomedismair3/Biofon/Biofon.java
@@ -10,8 +10,8 @@ import java.util.Arrays;
 
 public class Biofon
 {
-    public static final  short vendorId=(short)0xfc58;
-    public static final  short productId=(short)0x0001;
+    public static final  int vendorId=0xfc58;
+    public static final  int productId=0x0001;
     private static final  byte IN_END_POINT=(byte)0x81;
     private static final  byte OUT_END_POINT=(byte)0x01;
